requirements.txt
annotated-types==0.7.0
anyio==4.6.0
certifi==2024.8.30
click==8.1.7
colorama==0.4.6
comtypes==1.4.7
distro==1.9.0
Faker==30.1.0
fastapi==0.115.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
jiter==0.6.0
numpy==2.1.1
openai==1.51.1
pandas==2.2.3
pydantic==2.9.2
pydantic_core==2.23.4
pypiwin32==223
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.12
pyttsx3==2.98
pytz==2024.2
pywin32==307
six==1.16.0
sniffio==1.3.1
starlette==0.38.6
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.2
uvicorn==0.31.0
